sources/dashboard/requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
boto3==1.34.69
requests==2.31.0
joblib==1.3.2
python-dotenv==1.0.1
scikit-learn==1.4.1
pillow==10.2.0
opencv-python==4.9.0.80
tensorflow==2.15.0
torch==2.2.1
transformers==4.38.2
sentencepiece==0.2.0
protobuf==4.25.3 